requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo
ipykernel
seaborn
matplotlib
plotly
scikit-learn==1.6.1
imbalanced-learn
mlflow
dagshub
evidently

# BigQuery dependencies
google-cloud-bigquery
pandas-gbq
db-dtypes

# GCS and Vertex AI dependencies
google-cloud-storage
google-cloud-aiplatform

# PySpark dependencies
pyspark
pyarrow
findspark
dagshub

##-e .